header tests stan/lang/ast/fun

+++ src/stan/lang/ast/fun/infer_type_indexing_def.hpp
@@ -1,21 +1,16 @@
 #ifndef STAN_LANG_AST_FUN_INFER_TYPE_INDEXING_DEF_HPP
 #define STAN_LANG_AST_FUN_INFER_TYPE_INDEXING_DEF_HPP
 
-#include <stan/lang/ast/type/bare_expr_type.hpp>
-#include <stan/lang/ast/type/double_type.hpp>
-#include <stan/lang/ast/type/ill_formed_type.hpp>
-#include <stan/lang/ast/type/matrix_type.hpp>
-#include <stan/lang/ast/type/row_vector_type.hpp>
-#include <stan/lang/ast/type/vector_type.hpp>
+#include <stan/lang/ast.hpp>
 
 namespace stan {
   namespace lang {
 
     bare_expr_type infer_type_indexing(const bare_expr_type& bare_type,
-                                       size_t num_index_dims) {
+                                       int num_index_dims) {
       
       if (num_index_dims == 0) return bare_type;
-      if (num_index_dims > static_cast<size_t>(bare_type.num_dims()))
+      if (num_index_dims > bare_type.num_dims())
         return ill_formed_type();
 
       bare_expr_type tmp = bare_type;
